Fix issue with loading HTML from URL.
The code path for loading a demo from the gallery was still being run
after we loaded from the URL, clobbering the URL data.  The gist
codepath only worked because the jsonp request made it so the gallery
demo load was then overwritten by the correct data.

Now we only parse and load from the demo if we actually need it.