Fix JavaScript parsing error in IE 10 and earlier
While Cesium doesn't work in IE10, we need to allow the JS to parse so that
an app can still function without Cesium.  In this case, Uint8ClampedArray
isn't supported in IE 10, so the arraySlice polyfill was throwing on load.
I cleaned up `arraySlide` and deduplicated rom TypedArray code into
FeatureDetection.

To Test, load Viewer in IE11 with the debugger open and set to IE10
compatibility mode.