Fix bugs in TimeIntervalCollection.
* Undid extraction of local variable for intervals[index] in `removeInterval` which was not done correctly.
* Several places compared JulianDates incorrectly using TimeInterval.equals.

I also made the structure of the code match our original C# source as closely as
possible, given the language differences.  Both of the above errors were a
result of the code diverging, in an attempt to clean it up.

I also added another unit test ported from our original source, and kept
the new unit tests added in #6398 (after fixing them to be correct).