Add support for the Pelias geocoder
[Pelias](https://pelias.io/) is an open source geocoder originally built
by Mapzen. This adds `PeliasGeocoderService` for use with the Geocoder
widget.

Because Pelias (and some other geocoders) differentiate between an
`autocomplete` request and a `search` request, I added an optional
parameter to `GeocoderService.geocode` to provide that information. The
widget will us auto-complete as you are typing and `search` once you hit
enter or select a value from the dropdown.