Proposed Pull Request Template
### Description

I'm using the PR template draft to create the PR for the template draft! The proposed markdown file will automatically populate the body of any new PR. 

I tried to keep information here to a minimum since we want developers to carefully go through this on every pull request. If it's someone's first PR, I have the link to the contributor's guide of all the PR guidelines to be aware of. Otherwise, the basic structure is a description, the related issue(s) and confirming all the items on the checklist.

I wanted the description section to work as the space to say anything about the PR, including perhaps preemptive justifications (like, "this function might be contentious but here's why I chose to do it" or "there's one test that's failing and I can't figure out why" or "I couldn't test on Safari because I don't have a Mac"). 

I also chose to ask people to strike the irrelevant items instead of just checking them or deleting them. So if you were submitting a typo fix, (or a PR like this) none of them apply. If you were submitting something that fixes a bug, and no new tests were needed, you would communicate that by striking through. This allows the reviewer to double check items that the developer skipped if the reviewer suspects it actually might need a test or might need a change to `CHANGES.md` etc. 

This is just a draft so I welcome any feedback or criticism! For reference, these were good examples of other repo's templates that I looked at:

* https://github.com/devspace/awesome-github-templates#rocket-templates-for-pull-requests
* https://embeddedartistry.com/blog/2017/8/4/a-github-pull-request-template-for-your-projects

Do we need an item for adding documentation? What should it say? Anything else missing? 

### Related Issues

Resolves #6167 

### Checklist 

_Mark completed items with by placing an `x` between the square brackets. If an item is not relevant, ~strike~ it instead of marking as complete._

- [ ] ~I've added public API changes to `CHANGES.md`~.
- [ ] ~I've run all the tests and confirmed that all are passing~.
- [ ] ~I've added tests for new code I wrote and confirmed 100% test coverage (see [testing guide](https://github.com/AnalyticalGraphicsInc/cesium/blob/master/Documentation/Contributors/TestingGuide/README.md#running-the-tests) )~
- [ ] ~I've run `npm run eslint` to confirm there are no style errors~.
- [ ] ~Confirmed that it runs in all browsers (by running `npm run test -- --browsers Firefox,Chrome,Edge,Safari`)~.