Treat frustum changes as camera changes in Scene#render
Before this change, the user could update the frustum but not the camera
by resizing the viewer. The Scene would continue to render even when
requestRenderMode was enabled because it recognized the frustum had
changed, but would not update the cached camera to mark the change as
handled. This commit changes cameraEqual (in Scene.js) to take into
account the camera frustum.

Fixes #6812