Fix normal calculation for CylinderGeometry
The calculated normals of the curved side of the cylinder are wrong when the bottom and top radius are not equal, as the code assumes that the normal is always at right angles to the cylinder axis. This commit handles the more general case where the radii at the ends of the cylinder are not equal.