Make Geocoder terrain-aware
Previously, if a GeocoderService returned a single point, the geocoder
would slam into the ground as close as possible. If the service returned
a rectangle, it would only work with when the rectangle height did not
include large elevation changes (and slam into the ground otherwise).

This change queries the 4 corners and center of the rectangle (if the
terrain supports it) and adjusts the geocoder final position height to
be above terrain.