Delay load google-earth-dbroot-parser
`google-earth-dbroot-parser.js` is a huge dependency (487 KB source, 202KB
minified).  It's also only needed if you are using a Google Earth server.

In order to avoid bloat and paying the penalty every time Cesium is loaded,
this change loads it on demand the first time it's needed.

I didn't use jsonp for this because there's no server to parse the query
and wrap the code in a callback. Instead we just load the script directly
into a unlikely to collide global variable.

The file gets minified and copied to ThirdParty, similar to what we were
already doing with the wasm file.

Also fixed a bug in buildCesiumViewer where it wasn't copying all the
necessary files to its own build output.

Also extracted `loadAndExecuteScript` out of Resource.js, but had to keep
a version of it in Resource.js because cleaning up test abuse of it is
a much larger change best for another PR. (I'll write up an issue).