Make computeFlyToLocationForRectangle always return a Cartographic
Not sure what I was thinking when I wrote
`computeFlyToLocationForRectangle`, but this changes it so it always
returns a `Cartographic`. Previously it would return a Rectangle if terrain
wasn't set or didn't have `availability` which just made the calling code
confusing for no reason and lead to #7143.

`computeFlyToLocationForRectangle` is private, so this isn't a breaking
change.

Fixes #7143