Don't use requirejs with minified Cesium.
While the improvements in #7514 worked, they were a little more convoluted
than they had to be and also caused some weird edge cases where loading
Cesium in eval'd Node.js code didn't work.

This change avoids using requirejs entirely when loading minified Cesium
in Node and adds a smokescreen test to travis that fails in master. This
also avoids the odd global workaround I had to put in place previously.

I did have to modify the UMD headers for NoSleep.js and purify.js because
their preference is to look for Node.js first and not use AMD if it's
found, this was the root cause of the failures we were seeing in eval'd
code.  The smokescreen test I added will fail again if the problem were
reintroduced.

So overall, this make Cesium a little more node-friendly and just makes
things cleaner from our end.