Don't send ion access token to non-ion servers
Some ion-hosted assets can still point to additional external servers as,
part of their dataset. We were still appending the ion access_token in
the Accept header to these servers, which is at best pointless but at worst
would cause some CORS requests to fail because the Accept header was not
allowed.