Make PolygonGraphics.hierarchy always produce a PolygonHierarchy.
For backwards compatibility we preserve the existing behavior of allowing an array of positions to be set.
This makes the behavior more closely match the documentation. The ability to set an array of positions
was undocumented, though some Sandcastle examples do this.

This removes the need for conversion in PolygonGeometryUpdater at the point where the data is used.