Get rid of defineSuite
Cesium traditionally had a custom `defineSuite` function that wrapped
both the `define` used for importing requrejs modules and the `describe`
used to define a jasmine test block.

In order to move to ES6, we need to get rid of this helper function and
just use separate`define`/`describe` calls. Then, when we move to ES6,
the describe calls will stay as-is and the define call will be replaced
with ES6 module imports.

Since we also overloaded describe calls for category support, this was
an easy change to make and has zero functional changes.

To make the diff easy, I left indentation as-is.  When we move to ES6,
the indentation will actually go back to being correct.