Automatically convert most AMD modules to ES6
Nothings works, but the modules are converted to the format we need them
in. Submitting this as one large separate change so that we can actual
diff and review all remaining changes in a meaningful way.

1. Wrote a new gulp task convertToModules to which converts all of the
non-thirdparty code in Source and Specs to ES6. We use default exports
for all of our modules and will eventually re-export them as named exports
in the combined Cesium.js. The task will eventually go away but keeping it
around until we know we don't need it.

2. Ran the script to actually convert the modules.

3. Removed all unused import statements since they are now ES6 errors.

4. Also includes latest gltf-pipeline as ES6, see https://github.com/AnalyticalGraphicsInc/gltf-pipeline/pull/498