Clarify license for mouse gesture icons
We received a question regarding licensing of flaticon.com icons that were
introduced way back in 2014 with #1621.

Turns out that website changed dramatically and the link no longer points
to the same place it used to. I updated the MD to specficially state these
files are CC BY 3.0 and included a web archive link instead of the current
website.