Fix Geocoder auto-complete suggestions when hosted inside Web Components.
Geocoder was relying on `Event.target` which does not work with the
Shadow DOM. This prevented suggestions from appearing inside of web
components. Detect if shadow DOM is being used and access the composedPath
to retrieve the real target if so.

Also remove unnecessary contains check for onInputEnd because we can
just listen subscribe to the container itself rather than the entire
document.