Use Authorization header for ion tokens
When it was first introduced, ion used the accept header to specify the
token to trick the browser into avoiding a CORS preflight request on each
tile. The spec has since changed to enforce a preflight request on long
accept headers because it could otherwise serve as a vector for denial of
service attacks. Since we are no longer avoiding the preflight, we should
use the cleaner and correct Authorization header.