Remove `access_token` from gist loading.
GitHub has deprecated use of `access_token` in query parameters and is email spamming us nonstop about it.
Gists can be read anonymously with a rate-limit of 60 requests per hour per client IP address.

Since we stopped using gists for Sandcastle example code two years ago, the gist codepath is only for extremely old examples, so the rate limit is unlikely to be an issue in practice.