Blur previous element when interacting with Cesium canvas.
Interacting with a canvas does not automatically blur the previously
focused element. This leads to unexpected interaction if the last element
was an input field. For example, clicking the mouse wheel could lead to
the value in the field changing unexpectedly. The solution is to blur
whatever has focus as soon as canvas interaction begins.