Add space to JSDoc attributes
I select text using double click a lot - it selects whole words so that I don't need to be that precise in selecting the beginning of a word and the end of it. The words are delimited by a space, not by visual look. If I try to select e.g. `Cesium.Cartographic.ZERO` in the Cesium docs using this method then it results in `staticconstantCesium.Cartographic.ZERO` being selected because there are no spaces after `static` and `constant`. This commit adds the spaces there.

Tested: jsdoc -c Tools/jsdoc/conf.json