Add Transforms.eastNorthUp
I need to keep the camera where it is and just point it to an entity I'm
about to track. `Entity.viewFrom` accepts east-north-up and I don't see
an obvious way how to compute it from the entity and camera positions.

This function does that in a dumb way - just create two points (one
south from the entity and one down below camera) and compute the
distance to them. Maybe it could be done easier with some matrix
operations but I don't understand them.