Fix bad picking tests
`toPickPrimitive` doesn't actually render the scene, just picks the scene
at the time it is called.  Because these tests were never actually calling
`render`, there was no guarantee that what was being picked was correct
depending on the order the tests were run in.  Switch to calling
`renderForSpecs` instead of `initializeFrame` to ensure that a render
happens.

Fixes #9009