Remove default Mapbox token, Default Mapbox layers in BaseLayerPicker
Mapbox no longer provides CesiumJS with a free access token for
demonstration purposes. So this change removes the default token and also
Mapbox related layers from the `BaseLayerPicker` defaults.

`accessToken` is now required when constructing `MapboxImageryProvider` or
`MapboxStyleImageryProvider`.

This only affects out-of-the-box behavior and applications using their own
Mapbox access token are unaffected.