Deprecate BingMapsApi
A long time ago, Cesium created instances of `BingMapsGeocoderService` and
`BigMapsImageryProvider` behind the scenes by default using a demo key.
We provided the `BingMapsApi` singleton to make it easy to set a different
key in one place.

We haven't shipped with a demo key in a long time and there's really no
good reason to maintain the `BingMapsApi` singleton at the library level.