Enable use of Cesium release package with Node.js
If you unzipped the Cesium release package and pointed to it in a Node.js
application, it would fail because of missing `index.cjs` and
`Build/package.json`. This includes those files in the release zip so
that everything works the same if you retrieved the package from the npm
registry.