Only queue billboards for update if they are dirty.
BillboardCollection.prototype._updateBillboard function queues billboards for update if their dirty flag is FALSE. This leads to iteration over the entire BillboardCollection whenever a new billboard is added. This is a huge performance hit when the number of billboards starts to grow into the tens of thousands within a single collection. 
If this change is incorrect for some reason, then I would recommend adding a comment to this IF check as it is extremely unintuitive why we would queue billboards for update if they are not dirty.