Scene.js: set camera frustum and offset from WebXR provided XRPose
WebXR provides information in the XRPose that allows the VR device to
define its IPD and projection matrix. Now by default if WebXR is
detected, we take this information as a base to reconfigure the
Cesium.Camera that is used for the renders (one for each eye).

Implemented abstractions/factorization that allow the same workflow if
either plain old WebVR is executed, or new WebXR is executed. This
code which is mostly processing and applying the WebXR pose to the
Cesium Camera should be implemented in a class of its own and avoid
Scene.js from getting too large.

Also, the code needs to be optimized, as the HMD and frustum
parameters are not expected to change too often, so recalculating
these transformations on every frame is overkill. We just need for
adjustments on the Camera frustum and IPD parameters to trigger
recalculations on the transformations (on an as-needed basis).

Scene.useWebXR property allows VR sessions to revert to old, non-WebXR
    methods. Enough configuration options related to VR have been
    created to justify a VR-specific option class. This property is
    mostly used for development purposes.

projectionMatrix2Frustum: set a provided PerspectiveOffCenterFrustum's
    parameters from a projection matrix. Should be moved to be a
    method of PerspectiveOffCenterFrustum.

convertPerspectiveOffCenterFrustum: utility that allows a
    PerspectiveFrustum to be inferred from a
    PerspectiveOffCenterFrustum, although the latter are more
    general than the former, so the conversion is not always 1:1.