Fix race condition in OctahedralProjectedCubeMapSpec.js
`raises error event when environment map fails to load.`
assumes that the errorEvent listener will always be fired before
`projection.update` returns true, but inspecting the code
shows that is clearly not the case. It's possible for an error
to be thrown, the update to return true, another tick to run
and then the error event is raised. This lead to a Jasmine
`afterAll` failure pointing to line 222 of this file on my machine
when running the tests.

Wrapping the event in a promise and waiting for both
`projection.update` to return true and the event to be
handled is the way to ensure no race conditions in the
spec itself.