Update Release Guide
The Cesium Workshop is out of date. The code does not work with the current version of CesiumJS, and will need code updates.

In the meantime, its useless to continue updating the version.

I'll also note that nobody reported the workshop issues, so it may not be worth maintaining the workshop going forward.