types: mark `options` param of ImageryLayer.fromProviderAsync as optional
As the examples show, it is perfectly fine to not pass in the options object. However, in my typescript project, if I don't pass
in the options, I get a type error. This is intended to fix that error.

This was brought up in https://community.cesium.com/t/error-when-initialize-imagerylayer-fromproviderasync-without-options/31665