Lazily builds picking octree
WIP - the goal is to only build octrees on tiles
that are picked against. Beyond saving memory and worker time,
this allows us to conservatively rebuild octrees when
vertical exaggeration changes.

TODO: build the octree in a worker task, and fallback to
the default pick strategy while it's being built.